+++ tests/core/test_state_handling.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from hotglue_tap_sdk.helpers import _state
+from hotglue_singer_sdk.helpers import _state
 
 
 @pytest.fixture